+++ CHANGES.rst
@@ -78,6 +78,7 @@ Unreleased
 -   Model classes and the ``db`` instance are available without imports in
     ``flask shell``. :issue:`1089`
 -   ``Pagination`` ``per_page`` cannot be 0. :issue:`1091`
+-   ``Pagination`` ``max_per_page`` defaults to 100. :issue:`1091`
 
 
 Version 2.5.1